send user to speedrun when image clicked
