+++ packages/nextjs/app/page.tsx
@@ -1,33 +1,15 @@
-import { getFrameMetadata } from "@coinbase/onchainkit";
-import type { Metadata } from "next";
+"use client";
 
-const frameMetadata = getFrameMetadata({
-  buttons: [
-    {
-      label: "⚡ Start",
-    },
-  ],
-  image: `https://speedrunethereum.com/thumbnail.png`,
-  post_url: `${process.env.NEXT_PUBLIC_BASE_URL}/api/frame?id=0`,
-});
-
-export const metadata: Metadata = {
-  title: "Speed Run Ethereum",
-  description: "Learn how to build on Ethereum; the superpowers and the gotchas.",
-  openGraph: {
-    title: "Speed Run Ethereum",
-    description: "Learn how to build on Ethereum; the superpowers and the gotchas.",
-    images: [`https://speedrunethereum.com/thumbnail.png`],
-  },
-  other: {
-    ...frameMetadata,
-  },
-};
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
 
 export default function Page() {
-  return (
-    <>
-      <h1>Speed Run Ethereum</h1>
-    </>
-  );
+  const router = useRouter();
+
+  useEffect(() => {
+    // Redirect users to the desired URL
+    router.push("https://speedrunethereum.com/");
+  }, [router]);
+
+  return null;
 }
